models move to setup
